+++ android/src/main/java/com/theoplayer/presentation/PresentationManager.kt
@@ -10,13 +10,14 @@ import android.content.pm.PackageManager
 import android.os.Build
 import android.view.View
 import android.view.ViewGroup
-import android.view.ViewParent
 import androidx.activity.ComponentActivity
 import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.WindowInsetsControllerCompat
 import androidx.core.view.children
 import androidx.lifecycle.Lifecycle
 import com.facebook.react.ReactRootView
+import com.facebook.react.bridge.Arguments
+import com.facebook.react.bridge.ReactContext
 import com.facebook.react.uimanager.ThemedReactContext
 import com.facebook.react.views.view.ReactViewGroup
 import com.theoplayer.BuildConfig
@@ -26,6 +27,8 @@ import com.theoplayer.ReactTHEOplayerView
 import com.theoplayer.android.api.error.ErrorCode
 import com.theoplayer.android.api.error.THEOplayerException
 import com.theoplayer.android.api.player.PresentationMode
+import com.theoplayer.util.findReactRootView
+import com.theoplayer.util.getClosestParentOfType
 
 const val IS_TRANSITION_INTO_PIP = "isTransitioningToPip"
 const val IS_IN_PIP_MODE = "isInPictureInPictureMode"
@@ -71,6 +74,22 @@ class PresentationManager(
         }
       }
     }
+
+    // Emit dimension updates when layout changes
+    // This makes sure a dimension update is sent after fullscreen immersive animations complete.
+    rootView?.addOnLayoutChangeListener { view, _, _, _, _, _, _, _, _ ->
+      val density = reactContext.resources.displayMetrics.density.toDouble()
+      val params = Arguments.createMap().apply {
+        putMap("window", Arguments.createMap().apply {
+          putDouble("width", (view.width) / density)
+          putDouble("height", (view.height) / density)
+        })
+      }
+      reactContext
+        .getJSModule(ReactContext.RCTDeviceEventEmitter::class.java)
+        .emit("didUpdateDimensions", params)
+    }
+
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
       supportsPip =
         reactContext.packageManager.hasSystemFeature(PackageManager.FEATURE_PICTURE_IN_PICTURE)
@@ -269,12 +288,7 @@ class PresentationManager(
     get() = viewCtx.playerView.getClosestParentOfType()
 
   private val rootView: ReactRootView?
-    get() {
-      val activity = reactContext.currentActivity ?: return null
-      // Try to search in parents and as a fallback option from root to bottom using depth-first order
-      return reactPlayerGroup?.getClosestParentOfType()
-        ?: (activity.window.decorView.rootView as? ViewGroup)?.getClosestParentOfType(false)
-    }
+    get() = findReactRootView( reactContext, reactPlayerGroup)
 
   private fun reparentPlayerToRoot() {
     reactPlayerGroup?.let { playerGroup ->
@@ -283,10 +297,12 @@ class PresentationManager(
 
         // Re-parent the playerViewGroup to the root node
         parent.removeView(playerGroup)
-        rootView?.addView(playerGroup)
+        rootView?.let { root ->
+          root.addView(playerGroup)
 
-        // Attach an observer that overrides the react-native lay-out and forces fullscreen.
-        fullScreenLayoutObserver.attach(playerGroup)
+          // Attach an observer that overrides the react-native lay-out and forces fullscreen.
+          fullScreenLayoutObserver.attach(playerGroup, root)
+        }
       }
     }
   }
@@ -332,35 +348,6 @@ class PresentationManager(
       // Optionally fully stop play-out
       if (viewCtx.backgroundAudioConfig.stopOnBackground) viewCtx.player.stop()
     }
-  }
-}
-
-inline fun <reified T : View> ViewGroup.getClosestParentOfType(upward: Boolean = true): T? {
-  if (upward) {
-    // Search in the parent views of `this` view up to the root
-    var parent: ViewParent? = parent
-    while (parent != null && parent !is T) {
-      parent = parent.parent
-    }
-    return parent as? T
-  } else {
-    // Search in the children collection.
-    val viewStack = ArrayDeque(children.toList())
-    // Use Stack/LIFO instead of recursion
-    while (viewStack.isNotEmpty()) {
-      when (val view = viewStack.removeAt(0)) {
-        is T -> {
-          return view
-        }
-
-        is ViewGroup -> {
-          // Filling LIFO with all children of the ViewGroup: depth-first order
-          viewStack.addAll(0, view.children.toList())
-        }
-      }
-    }
-    // Found nothing
-    return null
   }
 }