+++ ios/THEOplayerRCTMainEventHandler.swift
@@ -7,7 +7,7 @@ public class THEOplayerRCTMainEventHandler {
     // MARK: Members
     private weak var player: THEOplayer?
     private weak var presentationModeContext: THEOplayerRCTPresentationModeContext?
-    private var loadedMetadataAndChapterTracksInfo: [[String:Any]] = []
+    private(set) var loadedMetadataAndChapterTracksInfo: [[String:Any]] = []
         
     // MARK: Events
     var onNativePlay: RCTDirectEventBlock?
@@ -274,9 +274,9 @@ public class THEOplayerRCTMainEventHandler {
         self.loadedMetadataListener = player.addEventListener(type: PlayerEventTypes.LOADED_META_DATA) { [weak self, weak player] event in
             if DEBUG_THEOPLAYER_EVENTS { PrintUtils.printLog(logText: "[NATIVE] Received LOADED_META_DATA event from THEOplayer") }
             if let wplayer = player,
-               let welf = self,
-               let forwardedLoadedMetadataEvent = self?.onNativeLoadedMetadata {
-              let metadata = THEOplayerRCTTrackMetadataAggregator.aggregateTrackMetadata(player: wplayer, metadataTracksInfo: welf.loadedMetadataAndChapterTracksInfo)
+               let self,
+               let forwardedLoadedMetadataEvent = self.onNativeLoadedMetadata {
+              let metadata = THEOplayerRCTTrackMetadataAggregator.aggregateTrackInfo(player: wplayer, metadataTracksInfo: self.loadedMetadataAndChapterTracksInfo)
                 forwardedLoadedMetadataEvent(metadata)
             }
         }